+++ pytorch_forecasting/models/base_model.py
@@ -2,6 +2,7 @@
 Timeseries models share a number of common characteristics. This module implements these in a common base class.
 """
 from collections import namedtuple
+import copy
 from copy import deepcopy
 import inspect
 from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
@@ -11,6 +12,7 @@ import numpy as np
 from numpy.lib.function_base import iterable
 import pandas as pd
 from pytorch_lightning import LightningModule
+from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities.parsing import AttributeDict, get_init_args
 import scipy.stats
 import torch
@@ -19,6 +21,7 @@ from torch.nn.utils import rnn
 from torch.optim.lr_scheduler import LambdaLR, ReduceLROnPlateau
 from torch.utils.data import DataLoader
 from tqdm.autonotebook import tqdm
+import yaml
 
 from pytorch_forecasting.data import TimeSeriesDataSet
 from pytorch_forecasting.data.encoders import EncoderNormalizer, GroupNormalizer, MultiNormalizer, NaNLabelEncoder
@@ -31,6 +34,7 @@ from pytorch_forecasting.metrics import (
     MultiHorizonMetric,
     MultiLoss,
     QuantileLoss,
+    convert_torchmetric_to_pytorch_forecasting_metric,
 )
 from pytorch_forecasting.optim import Ranger
 from pytorch_forecasting.utils import (
@@ -118,6 +122,15 @@ def _concatenate_output(
     return output_cat
 
 
+STAGE_STATES = {
+    RunningStage.TRAINING: "train",
+    RunningStage.VALIDATING: "val",
+    RunningStage.TESTING: "test",
+    RunningStage.PREDICTING: "predict",
+    RunningStage.SANITY_CHECKING: "sanity_check",
+}
+
+
 class BaseModel(LightningModule):
     """
     BaseModel from which new timeseries models should inherit from.
@@ -167,6 +180,8 @@ class BaseModel(LightningModule):
                     return self.to_network_output(prediction=prediction)
 
     """
+
+    CHECKPOINT_HYPER_PARAMS_SPECIAL_KEY = "__special_save__"
 
     def __init__(
         self,
@@ -220,7 +235,9 @@ class BaseModel(LightningModule):
         # update hparams
         frame = inspect.currentframe()
         init_args = get_init_args(frame)
-        self.save_hyperparameters({name: val for name, val in init_args.items() if name not in self.hparams})
+        self.save_hyperparameters(
+            {name: val for name, val in init_args.items() if name not in self.hparams and name not in ["self"]}
+        )
 
         # update log interval if not defined
         if self.hparams.log_val_interval is None:
@@ -228,15 +245,42 @@ class BaseModel(LightningModule):
 
         if not hasattr(self, "loss"):
             if isinstance(loss, (tuple, list)):
-                self.loss = MultiLoss(metrics=loss)
+                self.loss = MultiLoss(metrics=[convert_torchmetric_to_pytorch_forecasting_metric(l) for l in loss])
             else:
-                self.loss = loss
+                self.loss = convert_torchmetric_to_pytorch_forecasting_metric(loss)
         if not hasattr(self, "logging_metrics"):
-            self.logging_metrics = nn.ModuleList([l for l in logging_metrics])
+            self.logging_metrics = nn.ModuleList(
+                [convert_torchmetric_to_pytorch_forecasting_metric(l) for l in logging_metrics]
+            )
         if not hasattr(self, "output_transformer"):
             self.output_transformer = output_transformer
         if not hasattr(self, "optimizer"):  # callables are removed from hyperparameters, so better to save them
-            self._optimizer = self.hparams.optimizer
+            self.optimizer = self.hparams.optimizer
+
+        # delete everything from hparams that cannot be serialized with yaml.dump
+        # which is particularly important for tensorboard logging
+        hparams_to_delete = []
+        for k, v in self.hparams.items():
+            try:
+                yaml.dump(v)
+            except:  # noqa
+                hparams_to_delete.append(k)
+                if not hasattr(self, k):
+                    setattr(self, k, v)
+
+        self.hparams_special = getattr(self, "hparams_special", [])
+        self.hparams_special.extend(hparams_to_delete)
+        for k in hparams_to_delete:
+            del self._hparams[k]
+            del self._hparams_initial[k]
+
+    @property
+    def current_stage(self) -> str:
+        """
+        Available inside lightning loops.
+        :return: current trainer stage. One of ["train", "val", "test", "predict", "sanity_check"]
+        """
+        return STAGE_STATES[self.trainer.state.stage]
 
     @property
     def n_targets(self) -> int:
@@ -348,6 +392,18 @@ class BaseModel(LightningModule):
         log.update(self.create_log(x, y, out, batch_idx))
         return log
 
+    def validation_epoch_end(self, outputs):
+        self.epoch_end(outputs)
+
+    def test_step(self, batch, batch_idx):
+        x, y = batch
+        log, out = self.step(x, y, batch_idx)
+        log.update(self.create_log(x, y, out, batch_idx))
+        return log
+
+    def test_epoch_end(self, outputs):
+        self.epoch_end(outputs)
+
     def create_log(
         self,
         x: Dict[str, torch.Tensor],
@@ -380,9 +436,6 @@ class BaseModel(LightningModule):
                 x, out, batch_idx, prediction_kwargs=prediction_kwargs, quantiles_kwargs=quantiles_kwargs
             )
         return {}
-
-    def validation_epoch_end(self, outputs):
-        self.epoch_end(outputs)
 
     def step(
         self, x: Dict[str, torch.Tensor], y: Tuple[torch.Tensor, torch.Tensor], batch_idx: int, **kwargs
@@ -477,7 +530,14 @@ class BaseModel(LightningModule):
             else:
                 loss = self.loss(prediction, y)
 
-        self.log(f"{['val', 'train'][self.training]}_loss", loss, on_step=self.training, on_epoch=True, prog_bar=True)
+        self.log(
+            f"{self.current_stage}_loss",
+            loss,
+            on_step=self.training,
+            on_epoch=True,
+            prog_bar=True,
+            batch_size=len(x["decoder_target"]),
+        )
         log = {"loss": loss, "n_samples": x["decoder_lengths"].size(0)}
         return log, out
 
@@ -525,10 +585,11 @@ class BaseModel(LightningModule):
                 else:
                     target_tag = ""
                 self.log(
-                    f"{target_tag}{['val', 'train'][self.training]}_{metric.name}",
+                    f"{target_tag}{self.current_stage}_{metric.name}",
                     loss_value,
                     on_step=self.training,
                     on_epoch=True,
+                    batch_size=len(x["decoder_target"]),
                 )
 
     def to_network_output(self, **results):
@@ -639,7 +700,7 @@ class BaseModel(LightningModule):
                 log_indices = [0]
             for idx in log_indices:
                 fig = self.plot_prediction(x, out, idx=idx, add_loss_to_title=True, **kwargs)
-                tag = f"{['Val', 'Train'][self.training]} prediction"
+                tag = f"{self.current_stage} prediction"
                 if self.training:
                     tag += f" of item {idx} in global batch {self.global_step}"
                 else:
@@ -844,17 +905,17 @@ class BaseModel(LightningModule):
             lr = lrs[0]
         else:
             lr = lrs
-        if callable(self._optimizer):
+        if callable(self.optimizer):
             try:
-                optimizer = self._optimizer(
+                optimizer = self.optimizer(
                     self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay, **optimizer_params
                 )
             except TypeError:  # in case there is no weight decay
-                optimizer = self._optimizer(self.parameters(), lr=lr, **optimizer_params)
+                optimizer = self.optimizer(self.parameters(), lr=lr, **optimizer_params)
         elif self.hparams.optimizer == "adam":
             optimizer = torch.optim.Adam(self.parameters(), lr=lr)
         elif self.hparams.optimizer == "adamw":
-            optimizer = torch.optim.AdamW(self.parameters(), lr=lr)
+            optimizer = torch.optim.AdamW(self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay)
         elif self.hparams.optimizer == "ranger":
             optimizer = Ranger(self.parameters(), lr=lr, weight_decay=self.hparams.weight_decay)
         elif self.hparams.optimizer == "sgd":
@@ -875,34 +936,31 @@ class BaseModel(LightningModule):
         if isinstance(lr, (list, tuple)):  # change for each epoch
             # normalize lrs
             lrs = np.array(lrs) / lrs[0]
-            schedulers = [
-                {
-                    "scheduler": LambdaLR(optimizer, lambda epoch: lrs[min(epoch, len(lrs) - 1)]),
-                    "interval": "epoch",
-                    "reduce_on_plateau": False,
-                    "frequency": 1,
-                }
-            ]
-        elif self.val_dataloader() is None:  # no schedule and no validation loss means no scheduling to be used
-            schedulers = []
+            scheduler_config = {
+                "scheduler": LambdaLR(optimizer, lambda epoch: lrs[min(epoch, len(lrs) - 1)]),
+                "interval": "epoch",
+                "frequency": 1,
+                "strict": False,
+            }
+        elif self.hparams.reduce_on_plateau_patience is None:
+            scheduler_config = {}
         else:  # find schedule based on validation loss
-            schedulers = [
-                {
-                    "scheduler": ReduceLROnPlateau(
-                        optimizer,
-                        mode="min",
-                        factor=0.1,
-                        patience=self.hparams.reduce_on_plateau_patience,
-                        cooldown=self.hparams.reduce_on_plateau_patience,
-                        min_lr=self.hparams.reduce_on_plateau_min_lr,
-                    ),
-                    "monitor": "val_loss",  # Default: val_loss
-                    "interval": "epoch",
-                    "reduce_on_plateau": True,
-                    "frequency": 1,
-                }
-            ]
-        return [optimizer], schedulers
+            scheduler_config = {
+                "scheduler": ReduceLROnPlateau(
+                    optimizer,
+                    mode="min",
+                    factor=0.2,
+                    patience=self.hparams.reduce_on_plateau_patience,
+                    cooldown=self.hparams.reduce_on_plateau_patience,
+                    min_lr=self.hparams.reduce_on_plateau_min_lr,
+                ),
+                "monitor": "val_loss",  # Default: val_loss
+                "interval": "epoch",
+                "frequency": 1,
+                "strict": False,
+            }
+
+        return {"optimizer": optimizer, "lr_scheduler": scheduler_config}
 
     @classmethod
     def from_dataset(cls, dataset: TimeSeriesDataSet, **kwargs) -> LightningModule:
@@ -936,6 +994,10 @@ class BaseModel(LightningModule):
         )  # add dataset parameters for making fast predictions
         # hyper parameters are passed as arguments directly and not as single dictionary
         checkpoint["hparams_name"] = "kwargs"
+        # save specials
+        checkpoint[self.CHECKPOINT_HYPER_PARAMS_SPECIAL_KEY] = {k: getattr(self, k) for k in self.hparams_special}
+        # add special hparams them back to save the hparams correctly for checkpoint
+        checkpoint[self.CHECKPOINT_HYPER_PARAMS_KEY].update(checkpoint[self.CHECKPOINT_HYPER_PARAMS_SPECIAL_KEY])
 
     @property
     def target_names(self) -> List[str]:
@@ -954,6 +1016,9 @@ class BaseModel(LightningModule):
 
     def on_load_checkpoint(self, checkpoint: Dict[str, Any]) -> None:
         self.dataset_parameters = checkpoint.get("dataset_parameters", None)
+        # load specials
+        for k, v in checkpoint[self.CHECKPOINT_HYPER_PARAMS_SPECIAL_KEY].items():
+            setattr(self, k, v)
 
     def to_prediction(self, out: Dict[str, Any], use_metric: bool = True, **kwargs):
         """
@@ -1598,7 +1663,7 @@ class BaseModelWithCovariates(BaseModel):
                 x = np.linspace(-data["std"], data["std"], bins)
                 # reversing normalization for group normalizer is not possible without sample level information
                 if not isinstance(scaler, (GroupNormalizer, EncoderNormalizer)):
-                    x = scaler.inverse_transform(x)
+                    x = scaler.inverse_transform(x.reshape(-1, 1)).reshape(-1)
                     ax.set_xlabel(f"Normalized {name}")
 
                 if len(x) > 0: